added notes init query and extracted user_db funcs
